Fix board not filling the whole render window
Using integer here caused miscalculation and board wasn't filling its whole window - there were black borders on the right and bottom.

+++ Student-Life-Simulator/src/Agent.h
@@ -22,6 +22,6 @@ public:
 
 	[[nodiscard]] sf::Vector2i getPosition() const;
 	[[nodiscard]] uint16_t getId() const;
-	void draw(sf::RenderWindow& window, const uint16_t& fieldSize, const bool& isSingle = true);
+	void draw(sf::RenderWindow& window, const float& fieldSize, const bool& isSingle = true);
 	virtual void move(const uint16_t& boardSize);
 };
